requirements.txt
astropy==4.2.1
attrs==21.2.0
boto3==1.17.85
botocore==1.20.85
certifi==2021.5.30
chardet==4.0.0
ephem==3.7.7.1
greenlet==1.1.0
idna==2.10
importlib-metadata==4.4.0
iniconfig==1.1.1
jmespath==0.10.0
numpy==1.20.3
packaging==20.9
Pillow==8.2.0
pluggy==0.13.1
#psycopg2==2.8.6
psycopg2-binary==2.8.6
py==1.10.0
pyerfa==2.0.0
pyparsing==2.4.7
pytest==6.2.4
pytest-env==0.6.2
python-dateutil==2.8.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
SQLAlchemy==1.4.17
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.5
zipp==3.4.1
